Create a time-varying dataset

Create a time-varying dataset

time_varying(
  x,
  specs,
  exposure,
  ...,
  time_units = c("days", "seconds"),
  id = "pat_id",
  n_cores = 1
)

check_tv_data(x, time_units, id)

check_tv_exposure(x, expected_ids, time_units, id)

check_tv_specs(specs, expected_features = NULL)

Arguments

x
A data.frame with four columns: <id>, "feature", "datetime", "value"

specs
a data.frame with four columns: "feature", "use_for_grid", "lookback_start", "lookback_end", "aggregation". See details below.

exposure
a data.frame with (at least) three columns: <id>, "exposure_start", "exposure_stop"

...
Other arguments (not used at this time)

time_units
What time units should be used? Seconds or days

id
The id to use. Default is "pat_id"

n_cores
Number of cores to use

expected_ids
A vector of expected ids based on the data.

expected_features
A vector of expected features based on the data.

Details

The defaults for specs are to use everything for the grid creation, and to set lookback_start=0, with a message in both cases. Currently supported aggregation functions include counting ("count" or "n"), last-value-carried forward ("last value" or "lvcf"), any/none ("any" or "binary"), time since ("time since" or "ts"), min/max/mean, and the special "event" (for which lookbacks are ignored).

The lookback window begins at row_start - lookback_end and ends at row_start - lookback_start. Passing NA to either lookback changes the corresponding window boundary to exposure_start.

Examples

  data(tv_example)
  time_varying(tv_example$data, tv_example$specs, tv_example$exposure,
               time_units = "days", id = "mcn")

#>   mcn exposure_start exposure_stop  row_start   row_stop albumin_lvcf
#> 1   1     2022-02-01    2022-02-08 2022-02-01 2022-02-06           15
#> 2   1     2022-02-01    2022-02-08 2022-02-06 2022-02-07           NA
#> 3   1     2022-02-01    2022-02-08 2022-02-07 2022-02-08           NA
#> 4   2     2022-02-01    2022-02-08 2022-02-01 2022-02-02           NA
#> 5   2     2022-02-01    2022-02-08 2022-02-02 2022-02-03           12
#> 6   2     2022-02-01    2022-02-08 2022-02-03 2022-02-08           12
#>   albumin_ts neurosurgery_count neuro note_any neuro appointment_count
#> 1         31                  1              0                       0
#> 2         NA                  1              1                       0
#> 3         NA                  1              1                       1
#> 4         NA                  0              0                       0
#> 5          0                  0              0                       0
#> 6          1                  0              0                       0
#>   neuro appointment_ts death_event
#> 1                   NA           0
#> 2                   NA           0
#> 3                    0           0
#> 4                   NA           0
#> 5                   NA           1
#> 6                   NA           0